NP-1715 Move project option to own component
